Moved auth.js from auth to routes folder
